DELTA 05.11 tarne
Versiooni uuendamise käigus ei ole vaja rakenduse repositooriumi tühjendada.

Selle versiooniga tulevad kaasa parandused järgmistele vigadele:

* Minu dokumentide loetelu avamisel viga (EDHS-351)
* PDFi genereerimine ebaõnnestub OpenOffice'i tõttu (EDHS-333)
* Failis valemite asendamine ebaõnnestub OpenOffice'i tõttu (EHDS-251)
* Listi lehekülje suuruse valiku probleemid (EDHS-283)
* Asenduse valikul tööülesannete nimekiri ei muutu (EDHS-324)
* Teavituse lingi kaudu dokumenti avades pannakse kuva erinevatest dokumentidest kokku (EDHS-335, EDHS-343)
* Sarja valikul ei kandu üle JP andmed Internet Exploreriga (EDHS-313)

NB! Koos DELTA versiooni uuendamisega peab uuendama ka MSO Service teenuse versiooni. Palun sealjuures järgida uuendatud paigaldusjuhendit (mso-service/doc/Paigaldusjuhend.docx).

git-svn-id: https://svn.smit/tarkvara/edhs/trunk@95508 5fddf92a-9921-11dd-8d32-b999f448c456

+++ src/common/source/java/ee/webmedia/alfresco/common/bootstrap/AbstractNodeUpdater.java
@@ -246,7 +246,7 @@ public abstract class AbstractNodeUpdater extends AbstractModuleComponent {
             double completedPercent = completedSize * 100 / ((double) totalSize);
             double lastDocsPerSec = i * 1000 / ((double) (endTime - startTime));
             double totalDocsPerSec = thisRunCompletedSize * 1000 / ((double) (endTime - thisRunStartTime));
-            int etaMinutes = (int) ((totalSize - completedSize) * (endTime - thisRunStartTime) / (thisRunCompletedSize * 60000));
+            int etaMinutes = (int) (((long)(totalSize - completedSize)) * (endTime - thisRunStartTime) / (long)(thisRunCompletedSize * 60000));
             i = 0;
             log.info(String.format("%s: %6.2f%% completed - %7d of %7d, %5.1f docs per second (last), %5.1f (total), ETA %d min", processName,
                     completedPercent, completedSize, totalSize, lastDocsPerSec, totalDocsPerSec, etaMinutes));
